Rename workspace queries module
